school/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
bidict==0.23.1
click==8.2.0
colorama==0.4.6
fastapi==0.115.12
frozenlist==1.6.0
h11==0.16.0
idna==3.10
multidict==6.4.3
propcache==0.3.1
pydantic==2.11.4
pydantic_core==2.33.2
python-engineio==4.12.1
python-socketio==5.13.0
simple-websocket==1.1.0
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
wsproto==1.2.0
yarl==1.20.0
python-jose[cryptography]
passlib[bcrypt]
fastapi-pagination
fastapi
uvicorn
fastapi-mail[all] 
python-dotenv
redis